backend/requirements.txt
fastapi
uvicorn
pandas
requests
google-generativeai
python-dotenv
Pillow
fpdf2
python-multipart
